+++ src/components/video/index.test.tsx
@@ -98,35 +98,21 @@ describe("Video component", () => {
       expect(SolidFns.getUrl).toHaveBeenCalledWith(mockThing, mockProperty);
     });
 
-    it("When getUrl returns null, should throw an error", () => {
+    it("When getUrl returns null, with an errorComponent, should render the error", () => {
       jest.spyOn(console, "error").mockImplementation(() => {});
 
       (SolidFns.getUrl as jest.Mock).mockReturnValueOnce(null);
-      expect(() =>
-        render(
-          <Video thing={mockThing} property={mockProperty} title={mockTitle} />
-        )
-      ).toThrowErrorMatchingSnapshot();
-      expect(SolidFns.getUrl).toHaveBeenCalled();
 
-      // eslint-disable-next-line no-console
-      (console.error as jest.Mock).mockRestore();
-    });
-
-    it("Should throw error if initial fetch fails, if onError is not passed", async () => {
-      jest.spyOn(console, "error").mockImplementation(() => {});
-      (SolidFns.getFile as jest.Mock).mockRejectedValueOnce("Error in fetch");
-
-      const { getByText } = render(
-        <ErrorBoundary fallbackRender={({ error }) => <div>{error}</div>}>
-          <Video thing={mockThing} property={mockProperty} />
-        </ErrorBoundary>
+      const { asFragment } = render(
+        <Video
+          thing={mockThing}
+          property={mockProperty}
+          errorComponent={({ error }) => <span>{error.toString()}</span>}
+        />
       );
 
-      await waitFor(() => expect(getByText("Error in fetch")).toBeDefined());
-
-      // eslint-disable-next-line no-console
-      (console.error as jest.Mock).mockRestore();
+      expect(SolidFns.getUrl).toHaveBeenCalled();
+      expect(asFragment()).toMatchSnapshot();
     });
 
     it("Should call onError if initial fetch fails, if it is passed", async () => {